upload first beta version, v1.1
upload of the first ever beta version

current version 1.1

+++ admin/requests/deleteusertag.php
@@ -6,15 +6,18 @@
     include("../../phpscripts/logs.php");
 
     if(isset($_POST["id"])){
-        // TODO: Make sure user has permissions to do this!
-        $usertag = get_usertag_by_id($_POST["id"]);
-        if($usertag->isdefault){
-            echo "error:cant delete default usertag! link!";
+        if(tag_has_permission(get_current_usertag(), "usertagpnl_create_usertag")){
+            $usertag = get_usertag_by_id($_POST["id"]);
+            if($usertag->isdefault){
+                echo "error:cant delete default usertag! link!";
+            }else{
+                delete_usertag($_POST["id"]);
+                kick_all_users_from_usertag($_POST["id"]);
+                create_log(get_current_account()->username . " deleted the usertag '$usertag->name'");
+                echo "success";
+            }
         }else{
-            delete_usertag($_POST["id"]);
-            kick_all_users_from_usertag($_POST["id"]);
-            create_log(get_current_account()->username . " deleted the usertag '$usertag->name'");
-            echo "success";
+
         }
     }else{
         echo "error:missing information";
